+++ examples/02_DynamicTriangle.cpp
@@ -37,7 +37,7 @@ static std::vector<Vector2f> initial_position =
 void GLExample02::onWindowSizeChanged(const float width, const float height)
 {
   // Make sure the viewport matches the new window dimensions.
-  glCheck(glViewport(0, 0, width, height));
+  glCheck(glViewport(0, 0, static_cast<int>(width), static_cast<int>(height)));
 }
 
 //------------------------------------------------------------------
